Jmmabanta/andy integration

+++ app/src/main/java/com/cmput301w23t09/qrhunter/profile/ProfileController.java
@@ -131,7 +131,7 @@ public class ProfileController {
                         assert players != null;
                         if (players.contains(playerID)) {
                           String hash = doc.getId();
-                          long score = (long) doc.get("score");
+                          Integer score = (int) (long) doc.get("score");
                           qrCodes.add(new QRCode(hash, null, null, score, null, null, null, null));
                         }
                       }
